+++ sealed_example/bin/sealed/legacy/data/simple/mixed.sealed.dart
@@ -17,13 +17,9 @@ part of 'mixed.dart';
 abstract class Apple {
   const Apple._internal();
 
-  @factory
-  static AppleHold /*!*/ hold({
+  factory Apple.hold({
     @required Banana /*?*/ banana,
-  }) =>
-      AppleHold(
-        banana: banana,
-      );
+  }) = AppleHold;
 
   bool /*!*/ isHold() => this is AppleHold;
 
@@ -136,13 +132,9 @@ class AppleHold extends Apple with EquatableMixin {
 abstract class Banana {
   const Banana._internal();
 
-  @factory
-  static BananaHold /*!*/ hold({
+  factory Banana.hold({
     @required Apple /*?*/ apple,
-  }) =>
-      BananaHold(
-        apple: apple,
-      );
+  }) = BananaHold;
 
   bool /*!*/ isHold() => this is BananaHold;
 
@@ -259,35 +251,20 @@ class BananaHold extends Banana with EquatableMixin {
 abstract class Coconut {
   const Coconut._internal();
 
-  @factory
-  static CoconutTest1 /*!*/ test1({
+  factory Coconut.test1({
     @required int /*?*/ x,
     @required double /*?*/ y,
-  }) =>
-      CoconutTest1(
-        x: x,
-        y: y,
-      );
+  }) = CoconutTest1;
 
-  @factory
-  static CoconutTest2 /*!*/ test2({
+  factory Coconut.test2({
     @required int /*?*/ x,
     @required double /*?*/ y,
-  }) =>
-      CoconutTest2(
-        x: x,
-        y: y,
-      );
+  }) = CoconutTest2;
 
-  @factory
-  static CoconutHold /*!*/ hold({
+  factory Coconut.hold({
     @required Apple /*?*/ apple,
     @required Banana /*?*/ banana,
-  }) =>
-      CoconutHold(
-        apple: apple,
-        banana: banana,
-      );
+  }) = CoconutHold;
 
   bool /*!*/ isTest1() => this is CoconutTest1;