+++ sealed_example/bin/sealed/nullsafe/data/simple/meta.sealed.dart
@@ -20,26 +20,16 @@ part of 'meta.dart';
 abstract class Weather {
   const Weather._internal();
 
-  @factory
-  static PrefixSunny sunny() => PrefixSunny();
+  factory Weather.sunny() = PrefixSunny;
 
-  @factory
-  static BadWeather rainy({
+  factory Weather.rainy({
     required int rain,
-  }) =>
-      BadWeather(
-        rain: rain,
-      );
+  }) = BadWeather;
 
-  @factory
-  static VeryBadWeather windy({
+  factory Weather.windy({
     required double velocity,
     required double? angle,
-  }) =>
-      VeryBadWeather(
-        velocity: velocity,
-        angle: angle,
-      );
+  }) = VeryBadWeather;
 
   bool isSunny() => this is PrefixSunny;
 
@@ -180,9 +170,6 @@ abstract class Weather {
 class PrefixSunny extends Weather with EquatableMixin {
   const PrefixSunny() : super._internal();
 
-  @factory
-  PrefixSunny copy() => PrefixSunny();
-
   @override
   String toString() => 'Weather.sunny()';
 
@@ -200,14 +187,6 @@ class BadWeather extends Weather with EquatableMixin {
   }) : super._internal();
 
   final int rain;
-
-  @factory
-  BadWeather copy({
-    int? rain,
-  }) =>
-      BadWeather(
-        rain: rain ?? this.rain,
-      );
 
   @override
   String toString() => 'Weather.rainy(rain: $rain)';