+++ sealed_example/bin/sealed/nullsafe/distinct/simple/weather.sealed.dart
@@ -20,26 +20,16 @@ part of 'weather.dart';
 abstract class Weather {
   const Weather._internal();
 
-  @factory
-  static WeatherSunny sunny() => WeatherSunny();
+  factory Weather.sunny() = WeatherSunny;
 
-  @factory
-  static WeatherRainy rainy({
+  factory Weather.rainy({
     required int rain,
-  }) =>
-      WeatherRainy(
-        rain: rain,
-      );
+  }) = WeatherRainy;
 
-  @factory
-  static WeatherWindy windy({
+  factory Weather.windy({
     required double velocity,
     required double? angle,
-  }) =>
-      WeatherWindy(
-        velocity: velocity,
-        angle: angle,
-      );
+  }) = WeatherWindy;
 
   bool isSunny() => this is WeatherSunny;
 
@@ -180,9 +170,6 @@ abstract class Weather {
 class WeatherSunny extends Weather {
   const WeatherSunny() : super._internal();
 
-  @factory
-  WeatherSunny copy() => WeatherSunny();
-
   @override
   String toString() => 'Weather.sunny()';
 
@@ -200,14 +187,6 @@ class WeatherRainy extends Weather {
   }) : super._internal();
 
   final int rain;
-
-  @factory
-  WeatherRainy copy({
-    int? rain,
-  }) =>
-      WeatherRainy(
-        rain: rain ?? this.rain,
-      );
 
   @override
   String toString() => 'Weather.rainy(rain: $rain)';