+++ super_enum_sealed_example/test/nullsafe/test_result.dart
@@ -6,24 +6,24 @@ import '../utils/type_utils.dart';
 void main() {
   group('class Result', () {
     test('initialization', () {
-      final a = Result.error(exception: 12);
+      final a = Error(exception: 12);
       expect(a.exception, equals(12));
       expect(a, hasType<Error<Object?>>());
 
-      final b = Result.success(data: 1.5);
+      final b = Result.success(data: 1.5).asSuccess();
       expect(b.data, equals(1.5));
       expect(b, hasType<Success<double>>());
 
-      final c = Result.success(data: 1);
+      final c = Success(data: 1);
       expect(c.data, equals(1));
       expect(c, hasType<Success<int>>());
     });
 
     test('equality', () {
-      final a = Result.success(data: 1);
-      final b = Result.success(data: 1);
-      final c = Result.success(data: 2);
-      final d = Result.success(data: 0.75);
+      final a = Success(data: 1);
+      final b = Success(data: 1);
+      final c = Success(data: 2);
+      final d = Success(data: 0.75);
 
       expect(a, equals(a));
       expect(a, equals(b));