+++ super_enum_sealed_example/test/nullsafe/test_weather.dart
@@ -5,19 +5,19 @@ import '../../bin/nullsafe/weather.dart';
 void main() {
   group('class Weather', () {
     test('initialization', () {
-      final w1 = Weather.windy(velocity: 1.5, angle: null);
+      final w1 = Windy(velocity: 1.5, angle: null);
       expect(w1.velocity, equals(1.5));
       expect(w1.angle, isNull);
 
-      final w2 = Weather.windy(velocity: 1.5, angle: 50);
+      final w2 = Windy(velocity: 1.5, angle: 50);
       expect(w2.velocity, equals(1.5));
       expect(w2.angle, equals(50));
     });
 
     test('equality', () {
-      final r1 = Weather.rainy(rain: 0);
-      final r2 = Weather.rainy(rain: 0);
-      final r3 = Weather.rainy(rain: 10);
+      final r1 = Rainy(rain: 0);
+      final r2 = Rainy(rain: 0);
+      final r3 = Rainy(rain: 10);
 
       expect(r1, equals(r1));
       expect(r2, equals(r1));