+++ sealed_generators/CHANGELOG.md
@@ -1,7 +1,12 @@
-## 1.0.0
+## 1.2.0
 
-- Initial version
+- Using factory instead of static methods
+- Disabled copy method generation
 
 ## 1.1.0
 
 - Decoupled super_enum APIs from dart_sealed
+
+## 1.0.0
+
+- Initial version